Extract normaliseStroke
